# src/api/swaggerV6.yaml
swagger: '2.0'
info:
  description: Governify Registry
  version: '6'
  title: Governify Registry
  termsOfService: www.isa.us.es
  contact:
    name: ISA Research Group
    url: www.isa.us.es
    email: isagroup.us@gmail.com
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
host: registry.governify.io
basePath: /api/v6
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /states:
    get:
      tags:
        - state-registry
      description: XXX
      operationId: statesGET
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/agreement_state'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateRegistry
    delete:
      tags:
        - state-registry
      description: XXX
      operationId: statesDELETE
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateRegistry
  '/states/{agreement}/filtered':
    get:
      tags:
        - state-registry
      description: XXX
      operationId: statesFilter
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/agreement_state'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateRegistry
...
definitions:
...
  agreement_state:
    type: object
...
  error_model:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
...